backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.1.4
python-multipart==0.0.6
